+++ test/integration/integration/kafka_consume_messages_test.clj
@@ -99,15 +99,15 @@
                      (catch ExceptionInfo ex
                        (ex-data ex)))))))
 
-(def ^:private system-test-disabled-dql-service-integration
+(def ^:private system-test-disabled-dlq-service-integration
   (component/system-map
     :config (component.config/new-config "resources/config_test_dead_letter_disabled.json" :test :json)
     :producer (component/using (component.producer/new-mock-producer) [:config])
     :consumer (component/using (component.consumer/new-mock-consumer topic-consumers) [:config :producer])))
 
 
-(s-test/deftest kafka-consumer-component-test-wrong-schema-dql-disabled
-  (let [system (component/start system-test-disabled-dql-service-integration)
+(s-test/deftest kafka-consumer-component-test-wrong-schema-dlq-disabled
+  (let [system (component/start system-test-disabled-dlq-service-integration)
         producer (component.helper/get-component-content :producer system)]
 
     (component.producer/produce! {:topic :consumer-topic-test