+++ CHANGELOG.md
@@ -9,8 +9,8 @@ of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- Added the `:dead-letter-integration-enabled` (`true`|`false`) config option to toggle the integration with a DQL
-  service (aka [Wraith-King](https://github.com/macielti/wraith-king)).
+- Added the `:dead-letter-queue-service-integration-enabled` (`true`|`false`) config option to toggle the integration
+  with a DLQ service (aka [Wraith-King](https://github.com/macielti/wraith-king)).
 
 ## [15.18.18] - 2022-07-23